requirements.txt
streamlit>=1.30.0
streamlit-authenticator>=0.3.1
PyYAML>=6.0.1
